requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<4.4
drf-spectacular>=0.27,<0.28
django-redis>=5.4,<6
django-debug-toolbar>=4.4,<4.5
django-extensions>=3.2,<3.3
psycopg[binary]==3.2.10
python-dotenv>=1.0,<1.1
Pillow>=11.0,<11.1
gunicorn>=21.2,<22
whitenoise>=6.6,<7
algoliasearch>=3.0,<4
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=24.2,<24.3
django-taggit>=6.1,<7
django-quill-editor>=0.1,<1
django-jazzmin>=2.6,<4
sentry-sdk[django]>=2.18,<3
